+++ lib/features/settings/ui/pages/flags_page.dart
@@ -105,6 +105,7 @@ class FlagsBody extends ConsumerStatefulWidget {
     resendAttachments,
     enableHabitsPageFlag,
     enableDashboardsPageFlag,
+    enableDailyOsPageFlag,
     enableEventsFlag,
     enableSessionRatingsFlag,
     enableProjectsFlag,
@@ -157,6 +158,8 @@ class _FlagsBodyState extends ConsumerState<FlagsBody> {
         return Icons.repeat_rounded;
       case enableDashboardsPageFlag:
         return Icons.dashboard_rounded;
+      case enableDailyOsPageFlag:
+        return Icons.today_outlined;
       case enableEventsFlag:
         return Icons.event_rounded;
       case enableSessionRatingsFlag:
@@ -204,6 +207,8 @@ class _FlagsBodyState extends ConsumerState<FlagsBody> {
         return context.messages.configFlagEnableHabitsPageDescription;
       case enableDashboardsPageFlag:
         return context.messages.configFlagEnableDashboardsPageDescription;
+      case enableDailyOsPageFlag:
+        return context.messages.configFlagEnableDailyOsDescription;
       case enableEventsFlag:
         return context.messages.configFlagEnableEventsDescription;
       case enableSessionRatingsFlag:
@@ -251,6 +256,8 @@ class _FlagsBodyState extends ConsumerState<FlagsBody> {
         return context.messages.configFlagEnableHabitsPage;
       case enableDashboardsPageFlag:
         return context.messages.configFlagEnableDashboardsPage;
+      case enableDailyOsPageFlag:
+        return context.messages.configFlagEnableDailyOs;
       case enableEventsFlag:
         return context.messages.configFlagEnableEvents;
       case enableSessionRatingsFlag: